LRCI-4191 Add Testray1 implementations of all Testray entity classes
