+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/TestHistoryMap.java
@@ -8,6 +8,7 @@ package com.liferay.jenkins.results.parser;
 import com.liferay.jenkins.results.parser.testray.TestrayBuild;
 import com.liferay.jenkins.results.parser.testray.TestrayCaseResult;
 import com.liferay.jenkins.results.parser.testray.TestrayCaseType;
+import com.liferay.jenkins.results.parser.testray.TestrayFactory;
 import com.liferay.jenkins.results.parser.testray.TestrayRoutine;
 import com.liferay.jenkins.results.parser.testray.TestrayRun;
 import com.liferay.jenkins.results.parser.testray.TestrayServer;
@@ -493,7 +494,7 @@ public class TestHistoryMap {
 				return _testrayRun;
 			}
 
-			_testrayRun = new TestrayRun(
+			_testrayRun = TestrayFactory.newTestrayRun(
 				getLatestTestrayBuild(), getBatchName(), new ArrayList<File>());
 
 			return _testrayRun;
