--- modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/testray/DefaultTestrayServer.java
@@ -1,17 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.jenkins.results.parser.testray;
-
-/**
- * @author Michael Hashimoto
- */
-public class DefaultTestrayServer extends BaseTestrayServer {
-
-	public DefaultTestrayServer(String urlString) {
-		super(urlString);
-	}
-
-}
