+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/testray/Testray1TestrayServer.java
@@ -1,0 +1,134 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.jenkins.results.parser.testray;
+
+import com.liferay.jenkins.results.parser.JenkinsResultsParserUtil;
+
+import java.io.IOException;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import org.json.JSONArray;
+import org.json.JSONObject;
+
+/**
+ * @author Michael Hashimoto
+ */
+public class Testray1TestrayServer extends TestrayServer {
+
+	@Override
+	public TestrayCaseType getTestrayCaseType(String testrayCaseTypeName) {
+		if (_testrayCaseTypes != null) {
+			return _testrayCaseTypes.get(testrayCaseTypeName);
+		}
+
+		_testrayCaseTypes = new HashMap<>();
+
+		try {
+			JSONObject jsonObject = new JSONObject(
+				requestGet("/home/-/testray/case_types.json"));
+
+			JSONArray dataJSONArray = jsonObject.getJSONArray("data");
+
+			for (int i = 0; i < dataJSONArray.length(); i++) {
+				JSONObject dataJSONObject = dataJSONArray.getJSONObject(i);
+
+				_testrayCaseTypes.put(
+					dataJSONObject.getString("name"),
+					TestrayFactory.newTestrayCaseType(this, dataJSONObject));
+			}
+		}
+		catch (IOException ioException) {
+			throw new RuntimeException(ioException);
+		}
+
+		return _testrayCaseTypes.get(testrayCaseTypeName);
+	}
+
+	@Override
+	public TestrayProject getTestrayProjectByID(long projectID) {
+		_initTestrayProjects();
+
+		return _testrayProjectsByID.get(projectID);
+	}
+
+	@Override
+	public TestrayProject getTestrayProjectByName(String projectName) {
+		_initTestrayProjects();
+
+		return _testrayProjectsByName.get(projectName);
+	}
+
+	@Override
+	public List<TestrayProject> getTestrayProjects() {
+		_initTestrayProjects();
+
+		return new ArrayList<>(_testrayProjectsByName.values());
+	}
+
+	protected Testray1TestrayServer(String urlString) {
+		super(urlString);
+	}
+
+	private synchronized void _initTestrayProjects() {
+		if ((_testrayProjectsByID != null) &&
+			(_testrayProjectsByName != null)) {
+
+			return;
+		}
+
+		_testrayProjectsByID = new HashMap<>();
+		_testrayProjectsByName = new HashMap<>();
+
+		int current = 1;
+
+		while (true) {
+			try {
+				String projectAPIURLPath = JenkinsResultsParserUtil.combine(
+					"/home/-/testray/projects.json?cur=",
+					String.valueOf(current), "&delta=", String.valueOf(_DELTA),
+					"&orderByCol=testrayProjectId");
+
+				JSONObject jsonObject = new JSONObject(
+					requestGet(projectAPIURLPath));
+
+				JSONArray dataJSONArray = jsonObject.getJSONArray("data");
+
+				if (dataJSONArray.length() == 0) {
+					break;
+				}
+
+				for (int i = 0; i < dataJSONArray.length(); i++) {
+					JSONObject dataJSONObject = dataJSONArray.getJSONObject(i);
+
+					TestrayProject testrayProject =
+						TestrayFactory.newTestrayProject(this, dataJSONObject);
+
+					_testrayProjectsByID.put(
+						testrayProject.getID(), testrayProject);
+					_testrayProjectsByName.put(
+						testrayProject.getName(), testrayProject);
+				}
+			}
+			catch (IOException ioException) {
+				throw new RuntimeException(ioException);
+			}
+			finally {
+				current++;
+			}
+		}
+	}
+
+	private static final int _DELTA = 50;
+
+	private Map<String, TestrayCaseType> _testrayCaseTypes;
+	private Map<Long, TestrayProject> _testrayProjectsByID;
+	private Map<String, TestrayProject> _testrayProjectsByName;
+
+}
