+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/testray/TestrayBuild.java
@@ -1,26 +1,18 @@
 /**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
 package com.liferay.jenkins.results.parser.testray;
 
-import com.liferay.jenkins.results.parser.BuildReportFactory;
-import com.liferay.jenkins.results.parser.JenkinsResultsParserUtil;
 import com.liferay.jenkins.results.parser.TopLevelBuildReport;
 
-import java.io.IOException;
-
-import java.net.MalformedURLException;
 import java.net.URL;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import org.json.JSONArray;
 import org.json.JSONObject;
 
 /**
@@ -28,18 +20,6 @@ import org.json.JSONObject;
  */
 public class TestrayBuild implements Comparable<TestrayBuild> {
 
-	public TestrayBuild(TestrayRoutine testrayRoutine, JSONObject jsonObject) {
-		_testrayRoutine = testrayRoutine;
-		_jsonObject = jsonObject;
-
-		_testrayProject = testrayRoutine.getTestrayProject();
-		_testrayServer = testrayRoutine.getTestrayServer();
-
-		_testrayProductVersion = _testrayProject.getTestrayProductVersionByID(
-			jsonObject.getLong("testrayProductVersionId"));
-	}
-
-	@Override
 	public int compareTo(TestrayBuild testrayBuild) {
 		if (testrayBuild == null) {
 			throw new NullPointerException("Testray build is null");
@@ -51,11 +31,15 @@ public class TestrayBuild implements Comparable<TestrayBuild> {
 	}
 
 	public String getDescription() {
-		return _jsonObject.getString("description");
+		return _jsonObject.optString("description");
 	}
 
 	public long getID() {
-		return _jsonObject.getLong("testrayBuildId");
+		return _jsonObject.getLong("id");
+	}
+
+	public JSONObject getJSONObject() {
+		return _jsonObject;
 	}
 
 	public String getName() {
@@ -63,8 +47,7 @@ public class TestrayBuild implements Comparable<TestrayBuild> {
 	}
 
 	public String getPortalBranch() {
-		Matcher matcher = _portalBranchPattern.matcher(
-			_jsonObject.optString("description"));
+		Matcher matcher = _portalBranchPattern.matcher(getDescription());
 
 		if (!matcher.find()) {
 			return null;
@@ -74,8 +57,7 @@ public class TestrayBuild implements Comparable<TestrayBuild> {
 	}
 
 	public String getPortalSHA() {
-		Matcher matcher = _portalSHAPattern.matcher(
-			_jsonObject.optString("description"));
+		Matcher matcher = _portalSHAPattern.matcher(getDescription());
 
 		if (!matcher.find()) {
 			return null;
@@ -85,120 +67,25 @@ public class TestrayBuild implements Comparable<TestrayBuild> {
 	}
 
 	public JSONObject getRunsJSONObject() {
-		if (_runsJSONObject != null) {
-			return _runsJSONObject;
-		}
-
-		TestrayServer testrayServer = getTestrayServer();
-
-		try {
-			_runsJSONObject = JenkinsResultsParserUtil.toJSONObject(
-				JenkinsResultsParserUtil.combine(
-					String.valueOf(testrayServer.getURL()),
-					"/home/-/testray/runs.json?delta=200&testrayBuildId=",
-					String.valueOf(getID())));
-		}
-		catch (IOException ioException) {
-		}
-
-		return _runsJSONObject;
+		return null;
 	}
 
 	public String getStartYearMonth() {
-		Matcher matcher = _getTestrayAttachmentURLMatcher();
-
-		if (matcher == null) {
-			return null;
-		}
-
-		return matcher.group("startYearMonth");
+		return null;
 	}
 
 	public List<TestrayCaseResult> getTestrayCaseResults() {
-		return getTestrayCaseResults(null, null);
+		return null;
 	}
 
 	public List<TestrayCaseResult> getTestrayCaseResults(
 		TestrayCaseType testrayCaseType, TestrayRun testrayRun) {
 
-		List<TestrayCaseResult> testrayCaseResults = new ArrayList<>();
-
-		TestrayServer testrayServer = getTestrayServer();
-
-		StringBuilder sb = new StringBuilder();
-
-		sb.append(testrayServer.getURL());
-		sb.append("/home/-/testray/case_results.json?delta=");
-		sb.append(_PAGE_DELTA);
-		sb.append("&orderByCol=status_sortable");
-		sb.append("&orderByType=asc");
-		sb.append("&resetCur=false");
-		sb.append("&testrayBuildId=");
-		sb.append(getID());
-
-		if (testrayCaseType != null) {
-			sb.append("&testrayCaseTypeId=");
-			sb.append(testrayCaseType.getID());
-		}
-
-		if (testrayRun != null) {
-			sb.append("&testrayRunId=");
-			sb.append(testrayRun.getRunID());
-		}
-
-		long previousTestrayCaseResultID = -1;
-
-		for (int page = 1; page < _PAGE_COUNT; page++) {
-			try {
-				String testrayCaseResultsURL = sb + "&cur=" + page;
-
-				System.out.println(testrayCaseResultsURL);
-
-				JSONObject jsonObject = JenkinsResultsParserUtil.toJSONObject(
-					testrayCaseResultsURL,
-					_testrayServer.getHTTPAuthorization());
-
-				JSONArray dataJSONArray = jsonObject.getJSONArray("data");
-
-				if (dataJSONArray.isEmpty()) {
-					break;
-				}
-
-				JSONObject firstDataJSONObject = dataJSONArray.getJSONObject(0);
-
-				if (Objects.equals(
-						firstDataJSONObject.optLong("testrayCaseResultId"),
-						previousTestrayCaseResultID)) {
-
-					break;
-				}
-
-				previousTestrayCaseResultID = firstDataJSONObject.getLong(
-					"testrayCaseResultId");
-
-				for (int i = 0; i < dataJSONArray.length(); i++) {
-					JSONObject dataJSONObject = dataJSONArray.getJSONObject(i);
-
-					TestrayCaseResult testrayCaseResult = new TestrayCaseResult(
-						this, dataJSONObject);
-
-					testrayCaseResults.add(testrayCaseResult);
-				}
-
-				if (dataJSONArray.length() < _PAGE_DELTA) {
-					break;
-				}
-			}
-			catch (Exception exception) {
-				exception.printStackTrace();
-			}
-		}
-
-		return testrayCaseResults;
+		return null;
 	}
 
 	public TestrayProductVersion getTestrayProductVersion() {
-		return _testrayProductVersion;
+		return null;
 	}
 
 	public TestrayProject getTestrayProject() {
@@ -214,220 +101,44 @@ public class TestrayBuild implements Comparable<TestrayBuild> {
 	}
 
 	public TopLevelBuildReport getTopLevelBuildReport() {
-		if (_topLevelBuildReport != null) {
-			return _topLevelBuildReport;
-		}
-
-		URL topLevelBuildReportURL = getTopLevelBuildReportURL();
-
-		if (topLevelBuildReportURL == null) {
-			return null;
-		}
-
-		_topLevelBuildReport = BuildReportFactory.newTopLevelBuildReport(this);
-
-		return _topLevelBuildReport;
+		return null;
 	}
 
 	public URL getTopLevelBuildReportURL() {
-		Matcher matcher = _getTestrayAttachmentURLMatcher();
-
-		if (matcher == null) {
-			return null;
-		}
-
-		try {
-			URL url = new URL(
-				JenkinsResultsParserUtil.combine(
-					"http://", matcher.group("topLevelMasterHostname"),
-					"/userContent/jobs/", matcher.group("topLevelJobName"),
-					"/builds/", matcher.group("topLevelBuildNumber"),
-					"/build-report.json.gz"));
-
-			if (JenkinsResultsParserUtil.exists(url)) {
-				return url;
-			}
-
-			url = new URL(matcher.group());
-
-			if (JenkinsResultsParserUtil.exists(url)) {
-				return url;
-			}
-
-			return null;
-		}
-		catch (MalformedURLException malformedURLException) {
-			throw new RuntimeException(malformedURLException);
-		}
+		return null;
 	}
 
 	public URL getTopLevelBuildURL() {
-		Matcher matcher = _getTestrayAttachmentURLMatcher();
-
-		if (matcher == null) {
-			return null;
-		}
-
-		try {
-			return new URL(
-				JenkinsResultsParserUtil.combine(
-					"https://", matcher.group("topLevelMasterHostname"),
-					".liferay.com/job/", matcher.group("topLevelJobName"), "/",
-					matcher.group("topLevelBuildNumber"), "/"));
-		}
-		catch (MalformedURLException malformedURLException) {
-			throw new RuntimeException(malformedURLException);
-		}
+		return null;
 	}
 
 	public TestrayCaseResult getTopLevelTestrayCaseResult() {
-		List<TestrayCaseResult> testrayCaseResults = getTestrayCaseResults(
-			_testrayServer.getTestrayCaseType("Batch"), null);
-
-		for (TestrayCaseResult testrayCaseResult : testrayCaseResults) {
-			if (!Objects.equals(
-					testrayCaseResult.getName(), "Top Level Build")) {
-
-				continue;
-			}
-
-			return testrayCaseResult;
-		}
-
 		return null;
 	}
 
 	public URL getURL() {
-		try {
-			return new URL(_jsonObject.getString("htmlURL"));
-		}
-		catch (MalformedURLException malformedURLException) {
-			throw new RuntimeException(malformedURLException);
-		}
-	}
-
-	protected TestrayBuild(URL testrayBuildURL) {
-		Matcher matcher = _testrayBuildURLPattern.matcher(
-			testrayBuildURL.toString());
-
-		if (!matcher.find()) {
-			throw new RuntimeException("Invalid Build URL " + testrayBuildURL);
-		}
-
-		String serverURL = matcher.group("serverURL");
-
-		_testrayServer = TestrayFactory.newTestrayServer(
-			matcher.group("serverURL"));
-
-		try {
-			JSONObject jsonObject = JenkinsResultsParserUtil.toJSONObject(
-				JenkinsResultsParserUtil.combine(
-					serverURL, "/home/-/testray/builds/",
-					matcher.group("buildID"), ".json"),
-				_testrayServer.getHTTPAuthorization());
-
-			_jsonObject = jsonObject.getJSONObject("data");
-
-			_testrayRoutine = TestrayFactory.newTestrayRoutine(
-				JenkinsResultsParserUtil.combine(
-					String.valueOf(_testrayServer.getURL()),
-					"/home/-/testray/builds?testrayRoutineId=",
-					_jsonObject.getString("testrayRoutineId")));
-
-			_testrayProject = _testrayRoutine.getTestrayProject();
-
-			_testrayProductVersion =
-				_testrayProject.getTestrayProductVersionByID(
-					_jsonObject.getLong("testrayProductVersionId"));
-		}
-		catch (IOException ioException) {
-			throw new RuntimeException(ioException);
-		}
-	}
-
-	private Matcher _getTestrayAttachmentURLMatcher() {
-		if (_testrayAttachmentURLMatcher != null) {
-			return _testrayAttachmentURLMatcher;
-		}
-
-		TestrayCaseResult topLevelTestrayCaseResult =
-			getTopLevelTestrayCaseResult();
-
-		if (topLevelTestrayCaseResult != null) {
-			for (TestrayAttachment testrayAttachment :
-					topLevelTestrayCaseResult.getTestrayAttachments()) {
-
-				Matcher testrayAttachmentURLMatcher =
-					_testrayAttachmentURLPattern.matcher(
-						String.valueOf(testrayAttachment.getURL()));
-
-				if (testrayAttachmentURLMatcher.find()) {
-					_testrayAttachmentURLMatcher = testrayAttachmentURLMatcher;
-
-					return _testrayAttachmentURLMatcher;
-				}
-			}
-		}
-
-		List<TestrayCaseResult> testrayCaseResults = getTestrayCaseResults();
-
-		if (testrayCaseResults.isEmpty()) {
-			return null;
-		}
-
-		int count = 0;
-
-		for (TestrayCaseResult testrayCaseResult : testrayCaseResults) {
-			count++;
-
-			if (count >= 5) {
-				break;
-			}
-
-			for (TestrayAttachment testrayAttachment :
-					testrayCaseResult.getTestrayAttachments()) {
-
-				Matcher testrayAttachmentURLMatcher =
-					_testrayAttachmentURLPattern.matcher(
-						String.valueOf(testrayAttachment.getURL()));
-
-				if (testrayAttachmentURLMatcher.find()) {
-					_testrayAttachmentURLMatcher = testrayAttachmentURLMatcher;
-
-					return _testrayAttachmentURLMatcher;
-				}
-			}
-		}
-
 		return null;
 	}
 
-	private static final int _PAGE_COUNT = 100;
+	protected TestrayBuild(
+		TestrayRoutine testrayRoutine, JSONObject jsonObject) {
 
-	private static final int _PAGE_DELTA = 200;
+		_testrayRoutine = testrayRoutine;
+
+		_testrayProject = testrayRoutine.getTestrayProject();
+		_testrayServer = testrayRoutine.getTestrayServer();
+
+		_jsonObject = jsonObject;
+	}
 
 	private static final Pattern _portalBranchPattern = Pattern.compile(
 		"Portal Branch: (?<portalBranch>[^;]+);");
 	private static final Pattern _portalSHAPattern = Pattern.compile(
 		"Portal SHA: (?<portalSHA>[^;]+);");
-	private static final Pattern _testrayAttachmentURLPattern = Pattern.compile(
-		JenkinsResultsParserUtil.combine(
-			"https://testray.liferay.com/reports/production/logs/",
-			"(?<startYearMonth>\\d{4}-\\d{2})/",
-			"(?<topLevelMasterHostname>test-\\d+-\\d+)/",
-			"(?<topLevelJobName>[^/]+)/(?<topLevelBuildNumber>\\d+)/.*"));
-	private static final Pattern _testrayBuildURLPattern = Pattern.compile(
-		JenkinsResultsParserUtil.combine(
-			"(?<serverURL>https://[^/]+)/home/-/testray/runs\\?",
-			"testrayBuildId=(?<buildID>\\d+)"));
 
 	private final JSONObject _jsonObject;
-	private JSONObject _runsJSONObject;
-	private Matcher _testrayAttachmentURLMatcher;
-	private final TestrayProductVersion _testrayProductVersion;
 	private final TestrayProject _testrayProject;
 	private final TestrayRoutine _testrayRoutine;
 	private final TestrayServer _testrayServer;
-	private TopLevelBuildReport _topLevelBuildReport;
 
 }
