+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/testray/TestrayProductVersion.java
@@ -1,13 +1,10 @@
 /**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
 package com.liferay.jenkins.results.parser.testray;
 
-import com.liferay.jenkins.results.parser.JenkinsResultsParserUtil;
-
-import java.net.MalformedURLException;
 import java.net.URL;
 
 import org.json.JSONObject;
@@ -17,29 +14,12 @@ import org.json.JSONObject;
  */
 public class TestrayProductVersion {
 
-	public TestrayProductVersion(
-		TestrayProject testrayProject, JSONObject jsonObject) {
-
-		_testrayProject = testrayProject;
-		_jsonObject = jsonObject;
-
-		_testrayServer = testrayProject.getTestrayServer();
-
-		String urlString = JenkinsResultsParserUtil.combine(
-			String.valueOf(_testrayServer.getURL()),
-			"/home/-/testray/product_versions?testrayProjectId=",
-			String.valueOf(testrayProject.getID()));
-
-		try {
-			_url = new URL(urlString);
-		}
-		catch (MalformedURLException malformedURLException) {
-			throw new RuntimeException(malformedURLException);
-		}
+	public long getID() {
+		return _jsonObject.getLong("id");
 	}
 
-	public long getID() {
-		return _jsonObject.getLong("testrayProductVersionId");
+	public JSONObject getJSONObject() {
+		return _jsonObject;
 	}
 
 	public String getName() {
@@ -55,12 +35,21 @@ public class TestrayProductVersion {
 	}
 
 	public URL getURL() {
-		return _url;
+		return null;
+	}
+
+	protected TestrayProductVersion(
+		TestrayProject testrayProject, JSONObject jsonObject) {
+
+		_testrayProject = testrayProject;
+
+		_testrayServer = testrayProject.getTestrayServer();
+
+		_jsonObject = jsonObject;
 	}
 
 	private final JSONObject _jsonObject;
 	private final TestrayProject _testrayProject;
 	private final TestrayServer _testrayServer;
-	private final URL _url;
 
 }
