+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/testray/TestrayRoutine.java
@@ -1,93 +1,21 @@
 /**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
 package com.liferay.jenkins.results.parser.testray;
 
-import com.liferay.jenkins.results.parser.JenkinsResultsParserUtil;
-
-import java.io.IOException;
-
-import java.net.MalformedURLException;
 import java.net.URL;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.Date;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.TreeMap;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
-
-import org.json.JSONArray;
 import org.json.JSONObject;
 
 /**
  * @author Michael Hashimoto
  */
 public class TestrayRoutine {
-
-	public TestrayRoutine(
-		TestrayProject testrayProject, JSONObject jsonObject) {
-
-		_testrayProject = testrayProject;
-		_jsonObject = jsonObject;
-
-		_testrayServer = testrayProject.getTestrayServer();
-
-		String urlString = JenkinsResultsParserUtil.combine(
-			String.valueOf(_testrayServer.getURL()),
-			"/home/-/testray/builds?testrayRoutineId=",
-			String.valueOf(getID()));
-
-		try {
-			_url = new URL(urlString);
-		}
-		catch (MalformedURLException malformedURLException) {
-			throw new RuntimeException(
-				"Invalid Testray project URL " + urlString,
-				malformedURLException);
-		}
-	}
-
-	public TestrayRoutine(URL testrayRoutineURL) {
-		Matcher matcher = _testrayRoutineURLPattern.matcher(
-			testrayRoutineURL.toString());
-
-		if (!matcher.find()) {
-			throw new RuntimeException(
-				"Invalid Routine URL " + testrayRoutineURL);
-		}
-
-		_url = testrayRoutineURL;
-
-		String serverURL = matcher.group("serverURL");
-
-		_testrayServer = TestrayFactory.newTestrayServer(
-			matcher.group("serverURL"));
-
-		try {
-			JSONObject jsonObject = JenkinsResultsParserUtil.toJSONObject(
-				JenkinsResultsParserUtil.combine(
-					serverURL, "/home/-/testray/routines/",
-					matcher.group("routineID"), ".json"),
-				_testrayServer.getHTTPAuthorization());
-
-			_jsonObject = jsonObject.getJSONObject("data");
-
-			_testrayProject = _testrayServer.getTestrayProjectByID(
-				Long.parseLong(_jsonObject.getString("testrayProjectId")));
-		}
-		catch (IOException ioException) {
-			throw new RuntimeException(ioException);
-		}
-	}
 
 	public TestrayBuild createTestrayBuild(
 		TestrayProductVersion testrayProductVersion, String buildName) {
@@ -100,75 +28,15 @@ public class TestrayRoutine {
 		TestrayProductVersion testrayProductVersion, String buildName,
 		Date buildDate, String buildDescription, String buildSHA) {
 
-		if (testrayProductVersion == null) {
-			throw new RuntimeException("Please set a Testray product version");
-		}
-
-		if (JenkinsResultsParserUtil.isNullOrEmpty(buildName)) {
-			throw new RuntimeException("Please set a Testray build name");
-		}
-
-		StringBuilder sb = new StringBuilder();
-
-		sb.append("name=");
-		sb.append(buildName);
-		sb.append("&testrayProductVersionId=");
-		sb.append(testrayProductVersion.getID());
-		sb.append("&testrayRoutineId=");
-		sb.append(getID());
-
-		if (buildDate != null) {
-			String buildDateString = JenkinsResultsParserUtil.toDateString(
-				buildDate, "MM-dd'T'HH:mm:ss.SSS'Z'", "America/Los_Angeles");
-
-			sb.append("&createDate=");
-			sb.append(buildDateString);
-			sb.append("&dueDate=");
-			sb.append(buildDateString);
-			sb.append("&modifiedDate=");
-			sb.append(buildDateString);
-		}
-
-		if (!JenkinsResultsParserUtil.isNullOrEmpty(buildDescription)) {
-			sb.append("&description=");
-			sb.append(buildDescription);
-		}
-
-		if (!JenkinsResultsParserUtil.isNullOrEmpty(buildSHA)) {
-			sb.append("&gitHash=");
-			sb.append(buildSHA);
-		}
-
-		String buildAddURL = JenkinsResultsParserUtil.combine(
-			String.valueOf(_testrayServer.getURL()),
-			"/web/guest/home/-/testray/builds/add.json");
-
-		try {
-			JSONObject jsonObject = JenkinsResultsParserUtil.toJSONObject(
-				buildAddURL, 2, 5, sb.toString(),
-				_testrayServer.getHTTPAuthorization());
-
-			if (jsonObject.has("data")) {
-				return new TestrayBuild(this, jsonObject.getJSONObject("data"));
-			}
-
-			String message = jsonObject.optString("message", "");
-
-			if (!message.equals("The build name already exists.")) {
-				throw new RuntimeException("Failed to create a Testray build");
-			}
-		}
-		catch (IOException ioException) {
-			if (_log.isDebugEnabled()) {
-				_log.debug(ioException);
-			}
-		}
-
-		return getTestrayBuildByName(buildName);
+		return null;
 	}
 
 	public long getID() {
-		return _jsonObject.getLong("testrayRoutineId");
+		return _jsonObject.getLong("id");
+	}
+
+	public JSONObject getJSONObject() {
+		return _jsonObject;
 	}
 
 	public String getName() {
@@ -176,191 +44,23 @@ public class TestrayRoutine {
 	}
 
 	public TestrayBuild getTestrayBuildByID(long buildID) {
-		if (_testrayBuildsByID.containsKey(buildID)) {
-			return _testrayBuildsByID.get(buildID);
-		}
-
-		String buildAPIURL = JenkinsResultsParserUtil.combine(
-			String.valueOf(_testrayServer.getURL()),
-			"/web/guest/home/-/testray/builds/view.json?id=",
-			String.valueOf(buildID));
-
-		try {
-			JSONObject jsonObject = JenkinsResultsParserUtil.toJSONObject(
-				buildAPIURL, true, _testrayServer.getHTTPAuthorization());
-
-			if (!jsonObject.has("data")) {
-				return null;
-			}
-
-			JSONObject dataJSONObject = jsonObject.getJSONObject("data");
-
-			TestrayBuild testrayBuild = new TestrayBuild(this, dataJSONObject);
-
-			_addToTestrayBuildMaps(testrayBuild);
-
-			return testrayBuild;
-		}
-		catch (IOException ioException) {
-			throw new RuntimeException(ioException);
-		}
+		return null;
 	}
 
 	public TestrayBuild getTestrayBuildByName(
 		String buildName, String... names) {
 
-		if (_testrayBuildsByName.containsKey(buildName)) {
-			return _testrayBuildsByName.get(buildName);
-		}
-
-		int current = 1;
-
-		StringBuilder sb = new StringBuilder();
-
-		for (String name : names) {
-			sb.append("&name=");
-			sb.append(JenkinsResultsParserUtil.fixURL(name));
-		}
-
-		while (true) {
-			try {
-				String buildAPIURL = JenkinsResultsParserUtil.combine(
-					String.valueOf(_testrayServer.getURL()),
-					"/home/-/testray/builds.json?cur=", String.valueOf(current),
-					"&delta=", String.valueOf(_DELTA), sb.toString(),
-					"&orderByCol=testrayBuildId&testrayRoutineId=",
-					String.valueOf(getID()));
-
-				JSONObject jsonObject = JenkinsResultsParserUtil.toJSONObject(
-					buildAPIURL, true, _testrayServer.getHTTPAuthorization());
-
-				JSONArray dataJSONArray = jsonObject.getJSONArray("data");
-
-				if (dataJSONArray.length() == 0) {
-					break;
-				}
-
-				for (int i = 0; i < dataJSONArray.length(); i++) {
-					JSONObject dataJSONObject = dataJSONArray.getJSONObject(i);
-
-					TestrayBuild testrayBuild = new TestrayBuild(
-						this, dataJSONObject);
-
-					_addToTestrayBuildMaps(testrayBuild);
-
-					if (_testrayBuildsByName.containsKey(buildName)) {
-						return _testrayBuildsByName.get(buildName);
-					}
-				}
-
-				current++;
-			}
-			catch (IOException ioException) {
-				throw new RuntimeException(ioException);
-			}
-		}
-
 		return null;
 	}
 
 	public List<TestrayBuild> getTestrayBuilds() {
-		return getTestrayBuilds(_DELTA);
+		return null;
 	}
 
 	public List<TestrayBuild> getTestrayBuilds(
 		int maxSize, String... nameFilters) {
 
-		int current = 1;
-
-		StringBuilder sb = new StringBuilder();
-
-		if ((nameFilters != null) && (nameFilters.length > 0)) {
-			for (String nameFilter : nameFilters) {
-				if (JenkinsResultsParserUtil.isNullOrEmpty(nameFilter)) {
-					continue;
-				}
-
-				sb.append("&name=");
-
-				if (nameFilter.contains("-")) {
-					sb.append("%22");
-					sb.append(nameFilter);
-					sb.append("%22");
-				}
-				else {
-					sb.append(nameFilter);
-				}
-			}
-		}
-
-		while ((current * _DELTA) <= maxSize) {
-			try {
-				String buildAPIURL = JenkinsResultsParserUtil.combine(
-					String.valueOf(_testrayServer.getURL()),
-					"/home/-/testray/builds.json?cur=", String.valueOf(current),
-					"&delta=", String.valueOf(_DELTA), sb.toString(),
-					"&orderByCol=testrayBuildId&testrayRoutineId=",
-					String.valueOf(getID()));
-
-				JSONObject jsonObject = JenkinsResultsParserUtil.toJSONObject(
-					buildAPIURL, true, _testrayServer.getHTTPAuthorization());
-
-				JSONArray dataJSONArray = jsonObject.getJSONArray("data");
-
-				if (dataJSONArray.length() == 0) {
-					break;
-				}
-
-				for (int i = 0; i < dataJSONArray.length(); i++) {
-					JSONObject dataJSONObject = dataJSONArray.getJSONObject(i);
-
-					TestrayBuild testrayBuild = new TestrayBuild(
-						this, dataJSONObject);
-
-					if (_testrayBuildsByID.containsKey(testrayBuild.getID())) {
-						break;
-					}
-
-					_addToTestrayBuildMaps(testrayBuild);
-				}
-			}
-			catch (IOException ioException) {
-				throw new RuntimeException(ioException);
-			}
-			finally {
-				current++;
-			}
-		}
-
-		List<TestrayBuild> testrayBuilds = new ArrayList<>();
-
-		for (TestrayBuild testrayBuild : _testrayBuildsByID.values()) {
-			String testrayBuildName = testrayBuild.getName();
-
-			if ((nameFilters != null) && (nameFilters.length > 0)) {
-				boolean matches = true;
-
-				for (String nameFilter : nameFilters) {
-					if (JenkinsResultsParserUtil.isNullOrEmpty(nameFilter) ||
-						testrayBuildName.contains(nameFilter)) {
-
-						continue;
-					}
-
-					matches = false;
-
-					break;
-				}
-
-				if (!matches) {
-					continue;
-				}
-			}
-
-			testrayBuilds.add(testrayBuild);
-		}
-
-		return testrayBuilds;
+		return null;
 	}
 
 	public TestrayProject getTestrayProject() {
@@ -375,27 +75,35 @@ public class TestrayRoutine {
 		return _url;
 	}
 
-	private void _addToTestrayBuildMaps(TestrayBuild testrayBuild) {
-		_testrayBuildsByID.put(testrayBuild.getID(), testrayBuild);
-		_testrayBuildsByName.put(testrayBuild.getName(), testrayBuild);
+	public void setJSONObject(JSONObject jsonObject) {
+		_jsonObject = jsonObject;
 	}
 
-	private static final int _DELTA = 200;
+	protected TestrayRoutine(
+		TestrayProject testrayProject, JSONObject jsonObject) {
 
-	private static final Log _log = LogFactory.getLog(TestrayRoutine.class);
+		_testrayProject = testrayProject;
 
-	private static final Pattern _testrayRoutineURLPattern = Pattern.compile(
-		JenkinsResultsParserUtil.combine(
-			"(?<serverURL>https://[^/]+)/home/-/testray/builds\\?",
-			"testrayRoutineId=(?<routineID>\\d+)"));
+		_testrayServer = testrayProject.getTestrayServer();
 
-	private final JSONObject _jsonObject;
-	private final Map<Long, TestrayBuild> _testrayBuildsByID = new TreeMap<>(
-		Collections.reverseOrder());
-	private final Map<String, TestrayBuild> _testrayBuildsByName =
-		new HashMap<>();
-	private final TestrayProject _testrayProject;
-	private final TestrayServer _testrayServer;
-	private final URL _url;
+		_jsonObject = jsonObject;
+	}
+
+	protected TestrayRoutine(URL url) {
+		_url = url;
+	}
+
+	protected void setTestrayProject(TestrayProject testrayProject) {
+		_testrayProject = testrayProject;
+	}
+
+	protected void setTestrayServer(TestrayServer testrayServer) {
+		_testrayServer = testrayServer;
+	}
+
+	private JSONObject _jsonObject;
+	private TestrayProject _testrayProject;
+	private TestrayServer _testrayServer;
+	private URL _url;
 
 }
